requirements.txt
#List all the python libraries 
#os
pathlib 
tensorflow 
numpy 
boto3
configparser
Pillow